border experiment
I reduced the border size and the border slice percentage. What I am
trying to do here is get rid of an area of the border which is not
covered by the designated border image (or at least, a bit of unwanted
background is visible behind part of this border image).