minor tag edits
I made minor edits to some of the tags such as adding the optional
closing slash to empty elements and removing superfluous "type" values.
I specified the language as English in the HTML tag and added the
encoding meta tag. I added a comment which applies to early versions of
IE and allows styling of unknown elements.

I also added a bit of JQuery which I hope will get the dropdown menu's
toggle button to show when the user has scrolled down triggering the
affix and then scrolled back up.