jQuery workaround
For the moment I have edited the JQuery so that mousing over the title
ensures that the menu button is visible. Clicking on a back to top
button should now do the same thing. This gets around the problem of the
menu button, having disappeared at the affix point, not reappearing when
scrolling back to the top of the page.