moved section ids to preceding elements
I moved the section ids which the menu links navigate to. Now instead of the ids being located in the titles themselves, they are located in the "back to top" notes which precede every section heading. Before I fixed this, there was a problem with the menu button in the affix position: after clicking a menu link, the section heading would be partly covered by the menu button.