removed feet icon from menu button at a certain width
I removed the feet icon from the menu button at a particular media query, reducing the button's height. This is a width where in the affix position, the button displays in the middle of the screen, so if the user navigated to a section through the menu the button would partly cover up the section heading.