fix CSS rules re: background
Change the CSS rules regarding the main background image to fix a scrolling bug. It had been set to "fixed" but had not rendered that way on a mobile device.

Also expand the background to fill the whole width of the screen and set Dr. Cranium to only appear when the screen is wide enough that his head will not be blocked.